Safari-611.1.10/LayoutTests/fast/events/window-onerror12-expected.txt

Test that window.onerror is called on window object for events created using document.createEvent("error"). Bug 8519.

Main frame window.onerror:
msg: [object Event]
url: undefined
line: undefined
column: undefined
error: undefined
DONE.
